test(prm): extract shared PRM fixtures to keep the FR-006 ratchet green
Reconciling the escalation tests to feed genuinely distinct episodes (the
#2134 semantics) added mock-setup boilerplate to two files that are
already over the 500-line cap, tripping the diff-scoped ratchet.

No test or assertion changed: expect() counts and test-block counts are
byte-identical to origin/main (integration 73/17, index 59/32). The
reduction is pure deduplication into helpers/, which is not *.test.ts and
so does not count toward the cap:

- createMockConfig / createMockPatternMatch / createMockSession were
  defined identically in both files; they now live in helpers/fixtures.ts.
- setupEscalatingRepetitionMocks was local to one describe block while a
  near-identical inline copy sat in another that could not reach it. It
  moves to helpers/fixtures.ts, taking _internals as a parameter so the
  helper stays a pure builder, and both sites call it.
- episodeAt / createTickingDetectPatterns in helpers/episodes.ts derive a
  fresh non-overlapping step range per tick, which is what "three strikes"
  means now that the ladder counts occurrences rather than detections.

Verified load-bearing rather than vacuous: collapsing episodeAt() to a
constant [1, 3] -- the pre-fix behaviour of re-reporting one episode --
fails 11 tests, then reverts clean.

integration.test.ts 979 -> 866 (baseline 928); index.test.ts 1123
(baseline 1213).
